taw: printer support without changing to i386 architecture
To avoid a crash and also support secure virtualization, we need the TAW to be
built for the amd64 architecture.  Include printing support without changing
the architecture to force in the Brother proprietary drivers.  We should be
using a printer with an open-source driver, anyway.

Testing Done:
Bug Number: SW-1972
Review URL:
Reviewed by:

Signed-off-by: Eli Sennesh <esennesh@leaflabs.com>