SW-2684: TVMs: rearrange directory structure for tvm/config properly
Thanks to the way we were arranging tvm/config and supplying contents to the
TVM's under-construction directory structure, our Debian Installer preseeding
file didn't appear in the right directory to be detected as such and apply.
Rearrange and consolidate the contents we supply to the TVM build to correct
this, allowing us to supply arbitrary modifications/contents to the tvm/config/
directory rather than just a skeleton home directory.

Testing Done: None yet.

Signed-off-by: Eli Sennesh <esennesh@leaflabs.com>